feat: move is_member to api/v1
