--- src/hooks/userHooks/useUserSubscription/index.ts
@@ -1,24 +1,0 @@
-import { useApi } from "hooks/useApi";
-import userSubscriptionApi from "services/user/userSubscriptionApi";
-import IsMember from "types/apiResponses/IsMember";
-
-function useUserSubscription() {
-  function userIsMember() {
-    const {
-      refetch,
-      isLoading,
-      data: isMember,
-    } = useApi<IsMember>({
-      key: "UserIsMember",
-      fetchMethod: () => userSubscriptionApi.getIsMember(),
-    });
-
-    return { isMember, refetch, isLoading };
-  }
-
-  return {
-    userIsMember,
-  };
-}
-
-export default useUserSubscription;
