+++ src/services/api/subscriptionApi/index.ts
@@ -1,4 +1,5 @@
 import { AxiosResponse } from "axios";
+import IsMember from "types/apiResponses/IsMember";
 
 import { apiGet, apiPut } from "..";
 
@@ -8,6 +9,8 @@ const subscriptionApi = {
 
   putCancelSubscription: (token: string): Promise<AxiosResponse<any>> =>
     apiPut(`subscriptions/cancel_subscription?token=${token}`, {}),
+
+  getIsMember: (): Promise<AxiosResponse<IsMember>> => apiGet("users/is_member"),
 };
 
 export default subscriptionApi;
