+++ src/services/user/index.ts
@@ -47,7 +47,7 @@ export function initializeApi({ url, headers }: InitializeApiProps) {
     async (error) => {
       const originalRequest = error.config;
       // eslint-disable-next-line no-underscore-dangle
-      if (error.response.status === 403 && !originalRequest._retry) {
+      if (error.response?.status === 403 && !originalRequest._retry) {
         // eslint-disable-next-line no-underscore-dangle
         originalRequest._retry = true;
         const newToken = await requestNewToken();
